Week1/requirements.txt
fastapi
uvicorn
requests
numpy
pandas
mysqlclient
